Enum postgres::ConnectError

pub enum ConnectError {
    InvalidUrl(String),
    MissingUser,
    DbError(DbError),
    MissingPassword,
    UnsupportedAuthentication,
    NoSslSupport,
    SslError(SslError),
    IoError(IoError),
    BadResponse,
}

Reasons a new Postgres connection could fail

Variants

InvalidUrl

The provided URL could not be parsed

MissingUser

The URL was missing a user

DbError

An error from the Postgres server itself

MissingPassword

A password was required but not provided in the URL

UnsupportedAuthentication

The Postgres server requested an authentication method not supported by the driver

NoSslSupport

The Postgres server does not support SSL encryption

SslError

There was an error initializing the SSL session

IoError

There was an error communicating with the server

BadResponse

The server sent an unexpected response