Enum postgres::Error

pub enum Error {
    DbError(DbError),
    IoError(IoError),
    StreamDesynchronized,
    WrongType(Type),
    InvalidColumn,
    WasNull,
    BadResponse,
    BadData,
}

An error encountered when communicating with the Postgres server

Variants

DbError

An error reported by the Postgres server

IoError

An error communicating with the Postgres server

StreamDesynchronized

The communication channel with the Postgres server has desynchronized due to an earlier communications error.

WrongType

An attempt was made to convert between incompatible Rust and Postgres types

InvalidColumn

An attempt was made to read from a column that does not exist

WasNull

A value was NULL but converted to a non-nullable Rust type

BadResponse

The server returned an unexpected response

BadData

The server provided data that the client could not parse