Enum postgres::Kind

pub enum Kind {
    Simple,
    Array(Type),
    Range(Type),
}

Represents the kind of a Postgres type.

Variants

Simple

A simple type like VARCHAR or INTEGER.

Array

An array type along with the type of its elements.

Range

A range type along with the type of its elements.