Module postgres::notification

Asynchronous notifications.

Structs

BlockingIter
An iterator over notifications which will block if none are pending.

Iter
A fallible iterator over pending notifications.

Notification
An asynchronous notification.

Notifications
Notifications from the Postgres backend.

TimeoutIter
An iterator over notifications which will block for a period of time if none are pending.