Module postgres::rows

Query result rows.

Structs

IntoIter
An owning iterator over Rows.

Iter
An iterator over Rows.

LazyRows
A lazily-loaded iterator over the resulting rows of a query.

Row
A single result row of a query.

Rows
The resulting rows of a query.

Traits

RowIndex
A trait implemented by types that can index into columns of a row.