Struct postgres::DbError

pub struct DbError {
    // some fields omitted
}

A Postgres error or notice.

Methods

impl DbError

fn severity(&self) -> &str

The field contents are ERROR, FATAL, or PANIC (in an error message), or WARNING, NOTICE, DEBUG, INFO, or LOG (in a notice message), or a localized translation of one of these.

fn code(&self) -> &SqlState

The SQLSTATE code for the error.

fn message(&self) -> &str

The primary human-readable error message. This should be accurate but terse (typically one line).

fn detail(&self) -> Option<&str>

An optional secondary error message carrying more detail about the problem. Might run to multiple lines.

fn hint(&self) -> Option<&str>

An optional suggestion what to do about the problem. This is intended to differ from Detail in that it offers advice (potentially inappropriate) rather than hard facts. Might run to multiple lines.

fn position(&self) -> Option<&ErrorPosition>

An optional error cursor position into either the original query string or an internally generated query.

fn where_(&self) -> Option<&str>

An indication of the context in which the error occurred. Presently this includes a call stack traceback of active procedural language functions and internally-generated queries. The trace is one entry per line, most recent first.

fn schema(&self) -> Option<&str>

If the error was associated with a specific database object, the name of the schema containing that object, if any. (PostgreSQL 9.3+)

fn table(&self) -> Option<&str>

If the error was associated with a specific table, the name of the table. (Refer to the schema name field for the name of the table's schema.) (PostgreSQL 9.3+)

fn column(&self) -> Option<&str>

If the error was associated with a specific table column, the name of the column. (Refer to the schema and table name fields to identify the table.) (PostgreSQL 9.3+)

fn datatype(&self) -> Option<&str>

If the error was associated with a specific data type, the name of the data type. (Refer to the schema name field for the name of the data type's schema.) (PostgreSQL 9.3+)

fn constraint(&self) -> Option<&str>

If the error was associated with a specific constraint, the name of the constraint. Refer to fields listed above for the associated table or domain. (For this purpose, indexes are treated as constraints, even if they weren't created with constraint syntax.) (PostgreSQL 9.3+)

fn file(&self) -> &str

The file name of the source-code location where the error was reported.

fn line(&self) -> u32

The line number of the source-code location where the error was reported.

fn routine(&self) -> &str

The name of the source-code routine reporting the error.