Struct postgres::Notifications

pub struct Notifications<'conn> {
    // some fields omitted
}

An iterator over asynchronous notifications

Methods

impl<'conn> Notifications<'conn>

fn next_block(&mut self) -> Result<Notification>

Returns the oldest pending notification.

If no notifications are pending, blocks until one arrives.

fn next_block_for(&mut self, timeout: Duration) -> Option<Result<Notification>>

Returns the oldest pending notification

If no notifications are pending, blocks for up to timeout time, after which None is returned.

Example

use std::old_io::{IoError, IoErrorKind};
use std::time::Duration;

use postgres::Error;

match conn.notifications().next_block_for(Duration::seconds(2)) {
    Some(Ok(notification)) => println!("notification: {}", notification.payload),
    Some(Err(e)) => println!("Error: {:?}", e),
    None => println!("Wait for notification timed out"),
}