Module postgres::types

Traits dealing with Postgres data types

Structs

Field
Information about a field of a composite type.

Other
Information about an unknown type.

SessionInfo
A structure providing information for conversion methods.

Slice

WasNull
An error indicating that a NULL Postgres value was passed to a FromSql implementation that does not support NULL values.

WrongType
An error indicating that a conversion was attempted between incompatible Rust and Postgres types.

Enums

IsNull
An enum representing the nullability of a Postgres value.

Kind
Represents the kind of a Postgres type.

Type
A Postgres type.

Traits

FromSql
A trait for types that can be created from a Postgres value.

ToSql
A trait for types that can be converted into Postgres values.

Type Definitions

Oid
A Postgres OID.