Trait postgres::types::RawFromSql

pub trait RawFromSql {
    fn raw_from_sql<R: Reader>(ty: &Type, raw: &mut R) -> Result<Self>;
}

A utility trait used by FromSql implementations

Required Methods

fn raw_from_sql<R: Reader>(ty: &Type, raw: &mut R) -> Result<Self>

Creates a new value of this type from a Reader of Postgres data.

It is the caller's responsibility to ensure that Postgres data of this type can be turned in to this type.