Trait postgres::types::ToSql

pub trait ToSql: Debug {
    fn to_sql<W: ?Sized>(&self, ty: &Type, out: &mut W, ctx: &SessionInfo) -> Result<IsNull> where Self: Sized, W: Write;
    fn accepts(ty: &Type) -> bool where Self: Sized;
    fn to_sql_checked(&self, ty: &Type, out: &mut Write, ctx: &SessionInfo) -> Result<IsNull>;
}

A trait for types that can be converted into Postgres values.

Types

The following implementations are provided by this crate, along with the corresponding Postgres types:

Rust type 	Postgres type(s)

bool 	BOOL

i8 	"char"

i16 	SMALLINT, SMALLSERIAL

i32 	INT, SERIAL

u32 	OID

i64 	BIGINT, BIGSERIAL

f32 	REAL

f64 	DOUBLE PRECISION

String 	VARCHAR, CHAR(n), TEXT, CITEXT

&str 	VARCHAR, CHAR(n), TEXT, CITEXT

Vec&lt;u8&gt; 	BYTEA

&[u8] 	BYTEA

HashMap&lt;String, Option&lt;String&gt;&gt; 	HSTORE

In addition, some implementations are provided for types in third party crates. These are disabled by default; to opt into one of these implementations, activate the Cargo feature corresponding to the crate's name. For example, the serde_json feature enables the implementation for the serde_json::Value type.

Rust type 	Postgres type(s)

serialize::json::Json 	JSON, JSONB

serde_json::Value 	JSON, JSONB

time::Timespec 	TIMESTAMP, TIMESTAMP WITH TIME ZONE

chrono::NaiveDateTime 	TIMESTAMP

chrono::DateTime&lt;UTC&gt; 	TIMESTAMP WITH TIME ZONE

chrono::DateTime&lt;Local&gt; 	TIMESTAMP WITH TIME ZONE

chrono::DateTime&lt;FixedOffset&gt; 	TIMESTAMP WITH TIME ZONE

chrono::NaiveDate 	DATE

chrono::NaiveTime 	TIME

uuid::Uuid 	UUID

bit_vec::BitVec 	BIT, VARBIT

eui48::MacAddress 	MACADDR

Nullability

In addition to the types listed above, ToSql is implemented for Option<T> where T implements ToSql. An Option<T> represents a nullable Postgres value.

Arrays

ToSql is implemented for Vec<T> and &[T] where T implements ToSql, and corresponds to one-dimentional Postgres arrays with an index offset of 0.

Required Methods

fn to_sql<W: ?Sized>(&self, ty: &Type, out: &mut W, ctx: &SessionInfo) -> Result<IsNull> where Self: Sized, W: Write

Converts the value of self into the binary format of the specified Postgres Type, writing it to out.

The caller of this method is responsible for ensuring that this type is compatible with the Postgres Type.

The return value indicates if this value should be represented as NULL. If this is the case, implementations must not write anything to out.

fn accepts(ty: &Type) -> bool where Self: Sized

Determines if a value of this type can be converted to the specified Postgres Type.

fn to_sql_checked(&self, ty: &Type, out: &mut Write, ctx: &SessionInfo) -> Result<IsNull>

An adaptor method used internally by Rust-Postgres.

All implementations of this method should be generated by the to_sql_checked!() macro.