Struct postgres::notification::Notifications

pub struct Notifications<'conn> { /* fields omitted */ }

Notifications from the Postgres backend.

Methods

impl<'conn> Notifications<'conn>

fn len(&self) -> usize

Returns the number of pending notifications.

fn is_empty(&self) -> bool

Determines if there are any pending notifications.

fn iter<'a>(&'a self) -> Iter<'a>

Returns a fallible iterator over pending notifications.

Note

This iterator may start returning Some after previously returning None if more notifications are received.

fn blocking_iter<'a>(&'a self) -> BlockingIter<'a>

Returns a fallible iterator over notifications that blocks until one is received if none are pending.

The iterator will never return None.

fn timeout_iter<'a>(&'a self, timeout: Duration) -> TimeoutIter<'a>

Returns a fallible iterator over notifications that blocks for a limited time waiting to receive one if none are pending.

Note

This iterator may start returning Some after previously returning None if more notifications are received.