Struct postgres::Transaction

pub struct Transaction<'a> { /* private fields */ }

A representation of a PostgreSQL database transaction.

Transactions will implicitly roll back by default when dropped. Use the commit method to commit the changes made in the transaction. Transactions can be nested, with inner transactions implemented via savepoints.

Implementations

impl<'a> Transaction<'a>

pub fn commit(self) -> Result<(), Error>

Consumes the transaction, committing all changes made within it.

pub fn rollback(self) -> Result<(), Error>

Rolls the transaction back, discarding all changes made within it.

This is equivalent to Transaction’s Drop implementation, but provides any error encountered to the caller.

pub fn prepare(&mut self, query: &str) -> Result<Statement, Error>

Like Client::prepare.

pub fn prepare_typed(
&mut self,
query: &str,
types: &[Type]
) -> Result<Statement, Error>

Like Client::prepare_typed.

pub fn execute<T>(
&mut self,
query: &T,
params: &[&(dyn ToSql + Sync)]
) -> Result<u64, Error> where
T: ?Sized + ToStatement,

Like Client::execute.

pub fn query<T>(
&mut self,
query: &T,
params: &[&(dyn ToSql + Sync)]
) -> Result<Vec<Row>, Error> where
T: ?Sized + ToStatement,

Like Client::query.

pub fn query_one<T>(
&mut self,
query: &T,
params: &[&(dyn ToSql + Sync)]
) -> Result<Row, Error> where
T: ?Sized + ToStatement,

Like Client::query_one.

pub fn query_opt<T>(
&mut self,
query: &T,
params: &[&(dyn ToSql + Sync)]
) -> Result<Option<Row>, Error> where
T: ?Sized + ToStatement,

Like Client::query_opt.

pub fn query_raw<T, P, I>(
&mut self,
query: &T,
params: I
) -> Result<RowIter<'_>, Error> where
T: ?Sized + ToStatement,
P: BorrowToSql,
I: IntoIterator<Item = P>,
I::IntoIter: ExactSizeIterator,

Like Client::query_raw.

pub fn bind<T>(
&mut self,
query: &T,
params: &[&(dyn ToSql + Sync)]
) -> Result<Portal, Error> where
T: ?Sized + ToStatement,

Binds parameters to a statement, creating a “portal”.

Portals can be used with the query_portal method to page through the results of a query without being forced to consume them all immediately.

Portals are automatically closed when the transaction they were created in is closed.

Panics

Panics if the number of parameters provided does not match the number expected.

pub fn query_portal(
&mut self,
portal: &Portal,
max_rows: i32
) -> Result<Vec<Row>, Error>

Continues execution of a portal, returning the next set of rows.

Unlike query, portals can be incrementally evaluated by limiting the number of rows returned in each call to query_portal. If the requested number is negative or 0, all remaining rows will be returned.

pub fn query_portal_raw(
&mut self,
portal: &Portal,
max_rows: i32
) -> Result<RowIter<'_>, Error>

The maximally flexible version of query_portal.

pub fn copy_in<T>(&mut self, query: &T) -> Result<CopyInWriter<'_>, Error> where
T: ?Sized + ToStatement,

Like Client::copy_in.

pub fn copy_out<T>(&mut self, query: &T) -> Result<CopyOutReader<'_>, Error> where
T: ?Sized + ToStatement,

Like Client::copy_out.

pub fn simple_query(
&mut self,
query: &str
) -> Result<Vec<SimpleQueryMessage>, Error>

Like Client::simple_query.

pub fn batch_execute(&mut self, query: &str) -> Result<(), Error>

Like Client::batch_execute.

pub fn cancel_token(&self) -> CancelToken

Like Client::cancel_token.

pub fn transaction(&mut self) -> Result<Transaction<'_>, Error>

Like Client::transaction, but creates a nested transaction via a savepoint.

pub fn savepoint<I>(&mut self, name: I) -> Result<Transaction<'_>, Error> where
I: Into<String>,

Like Client::transaction, but creates a nested transaction via a savepoint with the specified name.