Struct postgres::transaction::Config

pub struct Config { /* fields omitted */ }

Configuration of a transaction.

Methods

impl Config

fn new() -> Config

Creates a new Config with no configuration overrides.

fn isolation_level(&mut self, isolation_level: IsolationLevel) -> &mut Config

Sets the isolation level of the configuration.

fn read_only(&mut self, read_only: bool) -> &mut Config

Sets the read-only property of a transaction.

If enabled, a transaction will be unable to modify any persistent database state.

fn deferrable(&mut self, deferrable: bool) -> &mut Config

Sets the deferrable property of a transaction.

If enabled in a read only, serializable transaction, the transaction may block when created, after which it will run without the normal overhead of a serializable transaction and will not be forced to roll back due to serialization failures.