Struct postgres::transaction::Transaction

pub struct Transaction<'conn> { /* fields omitted */ }

A transaction on a database connection.

The transaction will roll back by default.

Methods

impl<'conn> Transaction<'conn>

fn prepare(&self, query: &str) -> Result<Statement<'conn>>

Like Connection::prepare.

fn prepare_cached(&self, query: &str) -> Result<Statement<'conn>>

Like Connection::prepare_cached.

Note

The statement will be cached for the duration of the connection, not just the duration of this transaction.

fn execute(&self, query: &str, params: &[&ToSql]) -> Result<u64>

Like Connection::execute.

fn query<'a>(&'a self, query: &str, params: &[&ToSql]) -> Result<Rows>

Like Connection::query.

fn batch_execute(&self, query: &str) -> Result<()>

Like Connection::batch_execute.

fn transaction<'a>(&'a self) -> Result<Transaction<'a>>

Like Connection::transaction, but creates a nested transaction via a savepoint.

Panics

Panics if there is an active nested transaction.

fn savepoint<'a>(&'a self, name: &str) -> Result<Transaction<'a>>

Like Connection::transaction, but creates a nested transaction via a savepoint with the specified name.

Panics

Panics if there is an active nested transaction.

fn connection(&self) -> &'conn Connection

Returns a reference to the Transaction's Connection.

fn is_active(&self) -> bool

Like Connection::is_active.

fn set_config(&self, config: &Config) -> Result<()>

Alters the configuration of the active transaction.

fn will_commit(&self) -> bool

Determines if the transaction is currently set to commit or roll back.

fn set_commit(&self)

Sets the transaction to commit at its completion.

fn set_rollback(&self)

Sets the transaction to roll back at its completion.

fn commit(self) -> Result<()>

A convenience method which consumes and commits a transaction.

fn finish(self) -> Result<()>

Consumes the transaction, commiting or rolling it back as appropriate.

Functionally equivalent to the Drop implementation of Transaction except that it returns any error to the caller.