Module postgres::types

Types.

Structs

Field
Information about a field of a composite type.

Type
A Postgres type.

WasNull
An error indicating that a NULL Postgres value was passed to a FromSql implementation that does not support NULL values.

WrongType
An error indicating that a conversion was attempted between incompatible Rust and Postgres types.

Enums

Date
A wrapper that can be used to represent infinity with Type::Date types.

IsNull
An enum representing the nullability of a Postgres value.

Kind
Represents the kind of a Postgres type.

Timestamp
A wrapper that can be used to represent infinity with Type::Timestamp and Type::Timestamptz types.

Constants

ABSTIME
ABSTIME - absolute, limited-range date and time (Unix system time)

ABSTIME_ARRAY
ABSTIME[]

ACLITEM
ACLITEM - access control list

ACLITEM_ARRAY
ACLITEM[]

ANY
ANY

ANYARRAY
ANYARRAY

ANYELEMENT
ANYELEMENT

ANYENUM
ANYENUM

ANYNONARRAY
ANYNONARRAY

ANY_RANGE
ANYRANGE

BIT
BIT - fixed-length bit string

BIT_ARRAY
BIT[]

BOOL
BOOL - boolean, 'true'/'false'

BOOL_ARRAY
BOOL[]

BOX
BOX - geometric box '(lower left,upper right)'

BOX_ARRAY
BOX[]

BPCHAR
BPCHAR - char(length), blank-padded string, fixed storage length

BPCHAR_ARRAY
BPCHAR[]

BYTEA
BYTEA - variable-length string, binary values escaped

BYTEA_ARRAY
BYTEA[]

CHAR
CHAR - single character

CHAR_ARRAY
CHAR[]

CID
CID - command identifier type, sequence in transaction id

CIDR
CIDR - network IP address/netmask, network address

CIDR_ARRAY
CIDR[]

CID_ARRAY
CID[]

CIRCLE
CIRCLE - geometric circle '(center,radius)'

CIRCLE_ARRAY
CIRCLE[]

CSTRING
CSTRING

CSTRING_ARRAY
CSTRING[]

DATE
DATE - date

DATE_ARRAY
DATE[]

DATE_RANGE
DATERANGE - range of dates

DATE_RANGE_ARRAY
DATERANGE[]

EVENT_TRIGGER
EVENT_TRIGGER

FDW_HANDLER
FDW_HANDLER

FLOAT4
FLOAT4 - single-precision floating point number, 4-byte storage

FLOAT8
FLOAT8 - double-precision floating point number, 8-byte storage

FLOAT4_ARRAY
FLOAT4[]

FLOAT8_ARRAY
FLOAT8[]

GTS_VECTOR
GTSVECTOR - GiST index internal text representation for text search

GTS_VECTOR_ARRAY
GTSVECTOR[]

INDEX_AM_HANDLER
INDEX_AM_HANDLER

INET
INET - IP address/netmask, host address, netmask optional

INET_ARRAY
INET[]

INT2
INT2 - -32 thousand to 32 thousand, 2-byte storage

INT4
INT4 - -2 billion to 2 billion integer, 4-byte storage

INT8
INT8 - ~18 digit integer, 8-byte storage

INT2_ARRAY
INT2[]

INT2_VECTOR
INT2VECTOR - array of int2, used in system tables

INT2_VECTOR_ARRAY
INT2VECTOR[]

INT4_ARRAY
INT4[]

INT4_RANGE
INT4RANGE - range of integers

INT4_RANGE_ARRAY
INT4RANGE[]

INT8_ARRAY
INT8[]

INT8_RANGE
INT8RANGE - range of bigints

INT8_RANGE_ARRAY
INT8RANGE[]

INTERNAL
INTERNAL

INTERVAL
INTERVAL - @ <number> <units>, time interval

INTERVAL_ARRAY
INTERVAL[]

JSON
JSON

JSONB
JSONB - Binary JSON

JSONB_ARRAY
JSONB[]

JSON_ARRAY
JSON[]

LANGUAGE_HANDLER
LANGUAGE_HANDLER

LINE
LINE - geometric line

LINE_ARRAY
LINE[]

LSEG
LSEG - geometric line segment '(pt1,pt2)'

LSEG_ARRAY
LSEG[]

MACADDR
MACADDR - XX:XX:XX:XX:XX:XX, MAC address

MACADDR8
MACADDR8 - XX:XX:XX:XX:XX:XX:XX:XX, MAC address

MACADDR8_ARRAY
MACADDR8[]

MACADDR_ARRAY
MACADDR[]

MONEY
MONEY - monetary amounts, $d,ddd.cc

MONEY_ARRAY
MONEY[]

NAME
NAME - 63-byte type for storing system identifiers

NAME_ARRAY
NAME[]

NUMERIC
NUMERIC - numeric(precision, decimal), arbitrary precision number

NUMERIC_ARRAY
NUMERIC[]

NUM_RANGE
NUMRANGE - range of numerics

NUM_RANGE_ARRAY
NUMRANGE[]

OID
OID - object identifier(oid), maximum 4 billion

OID_ARRAY
OID[]

OID_VECTOR
OIDVECTOR - array of oids, used in system tables

OID_VECTOR_ARRAY
OIDVECTOR[]

OPAQUE
OPAQUE

PATH
PATH - geometric path '(pt1,...)'

PATH_ARRAY
PATH[]

PG_DDL_COMMAND
PG_DDL_COMMAND - internal type for passing CollectedCommand

PG_DEPENDENCIES
PG_DEPENDENCIES - multivariate dependencies

PG_LSN
PG_LSN - PostgreSQL LSN datatype

PG_LSN_ARRAY
PG_LSN[]

PG_NDISTINCT
PG_NDISTINCT - multivariate ndistinct coefficients

PG_NODE_TREE
PG_NODE_TREE - string representing an internal node tree

POINT
POINT - geometric point '(x, y)'

POINT_ARRAY
POINT[]

POLYGON
POLYGON - geometric polygon '(pt1,...)'

POLYGON_ARRAY
POLYGON[]

RECORD
RECORD

RECORD_ARRAY
RECORD[]

REFCURSOR
REFCURSOR - reference to cursor (portal name)

REFCURSOR_ARRAY
REFCURSOR[]

REGCLASS
REGCLASS - registered class

REGCLASS_ARRAY
REGCLASS[]

REGCONFIG
REGCONFIG - registered text search configuration

REGCONFIG_ARRAY
REGCONFIG[]

REGDICTIONARY
REGDICTIONARY - registered text search dictionary

REGDICTIONARY_ARRAY
REGDICTIONARY[]

REGNAMESPACE
REGNAMESPACE - registered namespace

REGNAMESPACE_ARRAY
REGNAMESPACE[]

REGOPER
REGOPER - registered operator

REGOPERATOR
REGOPERATOR - registered operator (with args)

REGOPERATOR_ARRAY
REGOPERATOR[]

REGOPER_ARRAY
REGOPER[]

REGPROC
REGPROC - registered procedure

REGPROCEDURE
REGPROCEDURE - registered procedure (with args)

REGPROCEDURE_ARRAY
REGPROCEDURE[]

REGPROC_ARRAY
REGPROC[]

REGROLE
REGROLE - registered role

REGROLE_ARRAY
REGROLE[]

REGTYPE
REGTYPE - registered type

REGTYPE_ARRAY
REGTYPE[]

RELTIME
RELTIME - relative, limited-range time interval (Unix delta time)

RELTIME_ARRAY
RELTIME[]

SMGR
SMGR - storage manager

TEXT
TEXT - variable-length string, no limit specified

TEXT_ARRAY
TEXT[]

TID
TID - (block, offset), physical location of tuple

TID_ARRAY
TID[]

TIME
TIME - time of day

TIMESTAMP
TIMESTAMP - date and time

TIMESTAMPTZ
TIMESTAMPTZ - date and time with time zone

TIMESTAMPTZ_ARRAY
TIMESTAMPTZ[]

TIMESTAMP_ARRAY
TIMESTAMP[]

TIMETZ
TIMETZ - time of day with time zone

TIMETZ_ARRAY
TIMETZ[]

TIME_ARRAY
TIME[]

TINTERVAL
TINTERVAL - (abstime,abstime), time interval

TINTERVAL_ARRAY
TINTERVAL[]

TRIGGER
TRIGGER

TSM_HANDLER
TSM_HANDLER

TSQUERY
TSQUERY - query representation for text search

TSQUERY_ARRAY
TSQUERY[]

TSTZ_RANGE
TSTZRANGE - range of timestamps with time zone

TSTZ_RANGE_ARRAY
TSTZRANGE[]

TS_RANGE
TSRANGE - range of timestamps without time zone

TS_RANGE_ARRAY
TSRANGE[]

TS_VECTOR
TSVECTOR - text representation for text search

TS_VECTOR_ARRAY
TSVECTOR[]

TXID_SNAPSHOT
TXID_SNAPSHOT - txid snapshot

TXID_SNAPSHOT_ARRAY
TXID_SNAPSHOT[]

UNKNOWN
UNKNOWN

UUID
UUID - UUID datatype

UUID_ARRAY
UUID[]

VARBIT
VARBIT - variable-length bit string

VARBIT_ARRAY
VARBIT[]

VARCHAR
VARCHAR - varchar(length), non-blank-padded string, variable storage length

VARCHAR_ARRAY
VARCHAR[]

VOID
VOID

XID
XID - transaction id

XID_ARRAY
XID[]

XML
XML - XML content

XML_ARRAY
XML[]

Traits

FromSql
A trait for types that can be created from a Postgres value.

ToSql
A trait for types that can be converted into Postgres values.

Type Definitions

Oid
A Postgres OID.