Module postgres::types

Traits dealing with Postgres data types

Reexports

pub use self::slice::Slice;

Structs

Other
Information about an unknown type.

Enums

IsNull
An enum representing the nullability of a Postgres value.

Kind
Represents the kind of a Postgres type.

Type
A Postgres type

Traits

FromSql
A trait for types that can be created from a Postgres value.

ToSql
A trait for types that can be converted into Postgres values.

Type Definitions

Oid
A Postgres OID